Updated and working version
Finally an updated and working version. Had to compare the cproj file to the original version and remove ASF related entries and files from project so that the machine would function properly.  I'm not sure exactly which setting did the trick but suffice it to say it's back up and running.